Added TESLA support to powered machines.

+++ src/main/java/exter/foundry/gui/GuiMetalAtomizer.java
@@ -92,7 +92,7 @@ public class GuiMetalAtomizer extends GuiFoundry
     }
 
     //Draw stored power bar.
-    int power = te_atomizer.getStoredFoundryEnergy() * POWER_HEIGHT / te_atomizer.getFoundryEnergyCapacity();
+    int power = (int)(te_atomizer.getStoredFoundryEnergy() * POWER_HEIGHT / te_atomizer.getFoundryEnergyCapacity());
     if(power > 0)
     {
       drawTexturedModalRect(window_x + POWER_X, window_y + POWER_Y + POWER_HEIGHT - power, POWER_OVERLAY_X, POWER_OVERLAY_Y + POWER_HEIGHT - power, POWER_WIDTH, power);
@@ -123,8 +123,8 @@ public class GuiMetalAtomizer extends GuiFoundry
     if(isPointInRegion(POWER_X,POWER_Y,POWER_WIDTH,POWER_HEIGHT,mousex,mousey))
     {
       List<String> currenttip = new ArrayList<String>();
-      int power = te_atomizer.getStoredFoundryEnergy();
-      int  max_power = te_atomizer.getFoundryEnergyCapacity();
+      long power = te_atomizer.getStoredFoundryEnergy();
+      long max_power = te_atomizer.getFoundryEnergyCapacity();
       currenttip.add("Energy: " + String.valueOf(power) + "/" + String.valueOf(max_power));
       drawHoveringText(currenttip, mousex, mousey, fontRendererObj);
     }
