+++ src/main/java/exter/foundry/gui/GuiMetalCaster.java
@@ -89,7 +89,7 @@ public class GuiMetalCaster extends GuiFoundry
     }
 
     //Draw stored power bar.
-    int power = te_caster.getStoredFoundryEnergy() * POWER_HEIGHT / te_caster.getFoundryEnergyCapacity();
+    int power = (int)(te_caster.getStoredFoundryEnergy() * POWER_HEIGHT / te_caster.getFoundryEnergyCapacity());
     if(power > 0)
     {
       drawTexturedModalRect(window_x + POWER_X, window_y + POWER_Y + POWER_HEIGHT - power, POWER_OVERLAY_X, POWER_OVERLAY_Y + POWER_HEIGHT - power, POWER_WIDTH, power);
@@ -112,8 +112,8 @@ public class GuiMetalCaster extends GuiFoundry
     if(isPointInRegion(POWER_X,POWER_Y,POWER_WIDTH,POWER_HEIGHT,mousex,mousey))
     {
       List<String> currenttip = new ArrayList<String>();
-      int power = te_caster.getStoredFoundryEnergy();
-      int  max_power = te_caster.getFoundryEnergyCapacity();
+      long power = te_caster.getStoredFoundryEnergy();
+      long max_power = te_caster.getFoundryEnergyCapacity();
       currenttip.add("Energy: " + String.valueOf(power) + "/" + String.valueOf(max_power));
       drawHoveringText(currenttip, mousex, mousey, fontRendererObj);
     }
